pyproject.toml
[project]
name = "board-pta-adapter"
version = "1.0.0"
description = "GPLT/XCPC Competition Leaderboard Adapter for PTA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[project.scripts]
gplt-gen = "adapter.cli.gplt:generate"
gplt-sync = "adapter.cli.gplt:synchronize"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adapter"]

[tool.ruff]
line-length = 120

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
